refactoring move constants to lisa level.
