+++ microsoft/testsuites/network/sriov.py
@@ -9,10 +9,12 @@ from lisa import (
     Environment,
     Logger,
     RemoteNode,
+    SkippedException,
     TcpConnetionException,
     TestCaseMetadata,
     TestSuite,
     TestSuiteMetadata,
+    constants,
     features,
     node_requirement,
     schema,
@@ -23,7 +25,6 @@ from lisa.features import NetworkInterface, SerialConsole
 from lisa.nic import NicInfo
 from lisa.sut_orchestrator import AZURE, READY
 from lisa.tools import Cat, Ethtool, InterruptInspector, Iperf3, Lspci
-from lisa.util import SkippedException, constants
 from lisa.util.shell import wait_tcp_port_ready
 from microsoft.testsuites.network.common import (
     cleanup_iperf3,
