+++ selftests/azure/test_prepare.py
@@ -7,11 +7,10 @@ from unittest.case import TestCase
 
 from azure.mgmt.compute.models import ResourceSku  # type: ignore
 
-from lisa import schema, search_space
+from lisa import LisaException, constants, schema, search_space
 from lisa.environment import Environment
 from lisa.sut_orchestrator import AZURE
 from lisa.sut_orchestrator.azure import common, platform_
-from lisa.util import LisaException, constants
 from lisa.util.logger import get_logger
